refactor: delete breadcrumbs
