+++ src/components/admin/app-sidebar/AppSidebar.vue
@@ -37,7 +37,7 @@ import VaSidebar
   from '../../../vuestic-theme/vuestic-components/va-sidebar/VaSidebar'
 import SidebarLink from './components/SidebarLink'
 import SidebarLinkGroup from './components/SidebarLinkGroup'
-import { navigationRoutes } from '../app-breadcrumbs/NavigationRoutes'
+import { navigationRoutes } from './NavigationRoutes'
 
 export default {
   name: 'app-sidebar',
