+++ src/i18n/cn.json
@@ -9,9 +9,6 @@
     "signUp": "注册",
     "termsOfUse": "使用条款"
   },
-  "breadcrumbs": {
-    "home": "首页"
-  },
   "buttons": {
     "advanced": "图标、组、下拉菜单",
     "button": "按钮",
