+++ src/i18n/en.json
@@ -17,9 +17,6 @@
     "text": "If you feel that it’s not right, please send us a message at ",
     "back_button": "Back to dashboard"
   },
-  "breadcrumbs": {
-    "home": "Home"
-  },
   "rating": {
     "singleIcon": "Single Icon",
     "twoIcons": "Two Icons",
