+++ src/i18n/pt_br.json
@@ -11,9 +11,6 @@
     "termsOfUse": "Termos de uso.",
     "welcome": "Bem vindo!"
   },
-  "breadcrumbs": {
-    "home": "Início"
-  },
   "buttons": {
     "advanced": "Icones, Grupos, Dropdowns",
     "button": "BOTÃO",
